test(client-s3): move unit tests to own folder
