+++ clients/client-s3/test/unit/flexibleChecksums.spec.ts
@@ -6,7 +6,7 @@ import chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 import { Readable } from "stream";
 
-import { S3 } from "../src/S3";
+import { S3 } from "../../src/S3";
 
 chai.use(chaiAsPromised);
 const { expect } = chai;
